backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.59.5
pandas==2.2.3
openpyxl==3.1.5
matplotlib==3.10.0
requests==2.32.5
pydantic==2.10.4
python-multipart==0.0.20
